+++ tests/compilers/test_ast.py
@@ -611,7 +611,7 @@ def test_compiler_macro_tag_try():
     """Check that try forms within defmacro/deftag are compiled correctly"""
     # https://github.com/hylang/hy/issues/1350
     can_compile("(defmacro foo [] (try None (except [] None)) `())")
-    can_compile("(deftag foo [] (try None (except [] None)) `())")
+    can_compile("(deftag foo [a] (try None (except [] None)) `())")
 
 
 def test_ast_good_yield_from():